Moving Snapchat helpers to separate file.
